+++ databricks-workspace-e2/main.tf
@@ -46,3 +46,13 @@ resource "databricks_mws_workspaces" "databricks" {
   storage_configuration_id = databricks_mws_storage_configurations.databricks.storage_configuration_id
   network_id               = databricks_mws_networks.networking.network_id
 }
+
+data "databricks_service_principal" "tfe_service_principal" {
+  application_id = var.tfe_service_principal_id
+}
+
+resource "databricks_mws_permission_assignment" "tfe_service_principal_admin" {
+  workspace_id = databricks_mws_workspaces.databricks.workspace_id
+  principal_id = data.databricks_service_principal.tfe_service_principal.id
+  permissions  = ["ADMIN"]
+}