+++ databricks-workspace-e2/variables.tf
@@ -56,3 +56,9 @@ variable "workspace_name_override" {
   default     = null
   description = "Override the workspace name. If not set, the workspace name will be set to the project, env, and service."
 }
+
+variable "tfe_service_principal_id" {
+  type        = string
+  default     = null
+  description = "ID for the service principal for Terraform Enterprise, if available. Will be added as a workspace admin"
+}